detector/requirements.txt
PyYAML>=6.0.1
psutil>=6.1.1
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
